Fixed FPS not measuring correctly. Added player & FPS func. Start StateManager.

+++ modules/TerminalWindow.h
@@ -28,7 +28,7 @@ namespace term_engine::modules {
 
     void ClearGlyphs();
 
-    int SetGlyph(const int& x_cell, const int& y_cell, const Uint32& character, const SDL_Color& foreground, const SDL_Color& background);
+    int SetGlyph(const int& x_cell, const int& y_cell, const Uint16& character, const SDL_Color& foreground, const SDL_Color& background);
     int SetGlyph(const int& x_cell, const int& y_cell, const utilities::Glyph& glyph);
     int SetGlyphs(std::function<int(std::vector<utilities::Glyph>&)> func);
     int FillGlyphs(std::function<utilities::Glyph()> generator);
